Move numba ir analysis code into new file
